update golang and remove cobra dependency
