+++ .gitignore
@@ -1,4 +1,4 @@
-sentences
+/sentences
 *.swp
 cmd/sentences/sentences
 binary
