+++ go.mod
@@ -1,5 +1,3 @@
 module github.com/neurosnap/sentences
 
-go 1.16
-
-require github.com/spf13/cobra v1.1.3 // indirect
+go 1.18
